+++ modules/core/04-channel/keeper/multihop_test.go
@@ -13,17 +13,14 @@ type MultihopTestSuite struct {
 
 	coordinator *ibctesting.Coordinator
 	paths       ibctesting.LinkedPaths
-
-	// A is the first endpoint in a multihop channel
-	A ibctesting.EndpointM
-	// Z is the first endpoint in a multihop channel
-	Z ibctesting.EndpointM
+	// multihop channel path
+	chanPath *ibctesting.PathM
 }
 
 // SetupTest is run before each test method in the suite
 func (s *MultihopTestSuite) SetupTest() {
 	s.coordinator, s.paths = ibctesting.CreateLinkedChains(&s.Suite, 5)
-	s.A, s.Z = ibctesting.NewEndpointMFromLinkedPaths(s.paths)
+	s.chanPath = s.paths.ToPathM()
 }
 
 // TestMultihopMultihopTestSuite runs all multihop related tests.
