+++ testing/coordinator_multihop.go
@@ -1,0 +1,6 @@
+package ibctesting
+
+// CoordinatorM coordinates testing set up for multihop channels.
+type CoordinatorM struct {
+	*Coordinator
+}
